# Conflicts:
#	src/aimlab/AIMLAB_main.cpp
#	src/aimlab/component.hpp

+++ src/aimlab/gameloop.hpp
@@ -73,7 +73,7 @@ public:
             decalsToSpawn.clear();
 
             // 객체 삭제 Swap And Pop 방식
-            for (int i = world3d.size() - 1; i > -1; --i)
+            for (size_t i = world3d.size() - 1; i > -1; --i)
             {
                 if (world3d[i]->state == Died)
                 {
